+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
  * Allow setting a condition when the `Cache` attribute should be applied
  * Deprecate passing a non-flat list of attributes to `Controller::setController()`
  * Deprecate the `Symfony\Component\HttpKernel\DependencyInjection\Extension` class, use the parent `Symfony\Component\DependencyInjection\Extension\Extension` class instead
+ * Allow using Expression or \Closure for `validationGroups` in `#[MapRequestPayload]` and `#[MapQueryString]`
 
 8.0
 ---